htr_gradio/pyproject.toml
[project]
name = "htr_gradio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gradio[mcp]>=5.33.1",
    "pillow>=11.2.1",
]
